+++ picasso/lib.py
@@ -1165,3 +1165,76 @@ def pick_areas_rectangle(
         (xs, ys), (xe, ye) = pick
         areas[i] = w * np.sqrt((xe - xs) ** 2 + (ye - ys) ** 2)
     return areas
+
+
+def plot_subclustering(
+    clustered_n_events: np.ndarray,
+    sparse_n_eveents: np.ndarray,
+    plot_path: str | list[str] = "",
+    return_fig: bool = False,
+) -> tuple[plt.Figure, plt.Axes] | tuple[None, None]:
+    """Plot the results of subclustering analysis, see
+    ``picasso.clusterer.test_subclustering``.
+
+    Parameters
+    ----------
+    clustered_n_events : np.ndarray
+        Number of events for clustered molecules.
+    sparse_n_eveents : np.ndarray
+        Number of events for sparse molecules.
+    plot_path : str or list of strs, optional
+        If provided, the plot is saved to this path. If a list of
+        strings is given, each is used to save a separate plot. Default
+        is "".
+    return_fig : bool, optional
+        If True, the figure and axes are returned. Default is False.
+
+    Returns
+    -------
+    fig, ax : (plt.Figure, plt.Axes) or (None, None)
+        Figure and axes if ``return_fig`` is True, otherwise
+        (None, None).
+    """
+    m_far = clustered_n_events.mean()
+    m_close = sparse_n_eveents.mean()
+    s_far = clustered_n_events.std()
+    s_close = sparse_n_eveents.std()
+
+    # create the plot
+    fig, ax1 = plt.subplots(1, figsize=(6, 3), constrained_layout=True)
+    min_bin, max_bin = np.percentile(clustered_n_events, [2.5, 97.5])
+    vals, counts = np.unique(clustered_n_events, return_counts=True)
+    ax1.bar(
+        vals,
+        counts,
+        width=0.8,
+        alpha=0.5,
+        label=f"Sparse {m_far:.1f} +/- {s_far:.1f}",
+        color="C0",
+    )
+    ax1.axvline(m_far, color="C0", linestyle="--")
+    vals, counts = np.unique(sparse_n_eveents, return_counts=True)
+    ax1.bar(
+        vals,
+        counts,
+        width=0.8,
+        alpha=0.5,
+        label=f"Clustered {m_close:.1f} +/- {s_close:.1f}",
+        color="C1",
+    )
+    ax1.axvline(m_close, color="C1", linestyle="--")
+    ax1.set_xlabel("Number of events")
+    ax1.set_ylabel("Counts")
+    ax1.set_xlim(min_bin - 1, max_bin + 1)
+    ax1.legend()
+    if len(plot_path):
+        if isinstance(plot_path, str):
+            plot_path = [plot_path]
+        for path in plot_path:
+            fig.savefig(path, dpi=300)
+
+    if return_fig:
+        return fig, ax1
+    else:
+        plt.close(fig)
+        return None, None